+++ tests/test_channel.py
@@ -72,7 +72,7 @@ def load_sampler(load_channel: Any):
 def test_channels(project: Project, rack: ChannelRack):
     assert len(rack) == project.channel_count
     assert rack.fit_to_steps is None
-    assert rack.height == 646
+    assert rack.height == 621
     assert [group.name for group in rack.groups] == ["Audio", "Generators", "Unsorted"]
     assert not rack.swing